Added credit card validation to the unit tests.  The Selenium tester is most likely going to be deleted and replaced with a JavaScript-based testing framework; this commit is just to save the work which was done.
The credit card validation added something like 2,000 tests, all of
which are necessary and cannot be removed.  Due to the legendarily
atrocious performance of the Selenium IE driver, this left Selenium no
longer a viable option.  I attempted to mitigate the performance issue
by implementing input simulation in JavaScript for the credit card
numbers.

This made IE performance tolerable, but it brought to light an apparent
flaw in Selenium.  It appears that Selenium opens a new loopback
connection for every interaction between client and server.  That's
every field interrogation, every remote JavaScript execution -
everything.  Evidence of this can be seen by doing a [netstat -n] as the
tests are running.  As execution time increases, the list of connections
grows until it finally reaches port 65535, at which point Selenium blows
up because the OS has hit its limit for open connections.

This problem has always been present in the unit tests, but it went
unnoticed for two reasons:

1.  There were far, far fewer tests before credit card validation was
added.

2.  Before input simulation was added in this revision, performance was
slow enough that older connections would time out as the tests were
running.  Since the Chrome driver is an order of magnitude faster than
the IE driver, it tears through the credit card tests, so there's not
enough time for the connections to time out.

As such, the unit tests are now broken, and Selenium is no longer a
viable option.  As mentioned at the start of this comment, this commit
is just to save the work which was done.  Now that I have a working
solution for simulating input with JavaScript, I'll need to look into
porting these tests to a Javascript-based testing framework.

To close with an observation which is no longer relevant, it appears I
might have broken page reloading in this revision - a hack which was
added due to an IE bug.